requirements.txt
google-cloud-bigquery
pandas
requests
dbt-bigquery
numpy